refactor: Delimiter 원시값 포장
